+++ src/app/core/components/path-not-found-component/path-not-found-component.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-path-not-found-component',
+  templateUrl: './path-not-found-component.component.html',
+  styleUrls: ['./path-not-found-component.component.scss'],
+  standalone: true,
+  imports: [CommonModule],
+})
+export class PathNotFoundComponentComponent {
+
+}