+++ src/app/products/components/index.ts
@@ -1,2 +1,3 @@
 export * from './product-component/product.component'
 export * from './product-list/product-list.component'
+export * from './product-view/product-view.component'